refactor(api): extract tool-profile and agent-template handlers into routes/agent_templates.rs
routes/system.rs continued to span 8 unrelated domains (#3749). Move the
tool-profile + agent-template sub-domain (list_profiles, get_profile,
list_agent_templates, get_agent_template, get_agent_template_toml plus
the validate_template_name helper and template_name_validation_tests)
into a sibling module routes/agent_templates.rs and merge it back into
system::router(). Public route paths under /profiles* and /templates*
are unchanged; glob re-export via routes::* keeps openapi.rs path
bindings working. The shared librefang_home() helper in system.rs is
promoted to pub(super) so the new module can reach it via super::system.
The lone in-tree consumer (providers.rs unit tests) is updated from
crate::routes::system::{get_profile,list_profiles} to
crate::routes::agent_templates::{...}.

system.rs goes from 5939 -> 5662 lines (-277); new agent_templates.rs is
314 lines.

Refs #3749 (2 of 8 sub-domain extractions -- remaining: audit_log,
tools, sessions, approvals, hooks, commands, batch_resolve).
